split layout.ts into multiple files, implement automatic discrimination (untested)
